BEMOSS_requirements.txt
blist==1.3.6
decorator==4.0.9
dnspython==1.12.0
django==1.11
django-filter==1.1
bacpypes==0.16.2
djangorestframework==3.5.4
futures==3.0.5
hammock==0.2.4
lxml==4.2.4
netifaces==0.10.4
networkx==1.11
nose==1.3.3
-e git+https://github.com/EnerNOC/oadr2-ven-python.git@d37d3e83008d243992021c4c20ae736283ac2199#egg=oadr2_ven
Pillow==2.9.0
pandas==0.18.1
psycopg2==2.7.5
pycrypto==2.6.1
pymodbus==1.2.0
#pyOpenSSL==0.13.1
scales==1.0.9
sleekxmpp==1.3.1
spyrk==0.0.3
neurio
tornado==4.4.1
tablib==0.11.2
wheel==0.24.0
python-jose==1.3.2
simplejson
gevent
guppy
sklearn==0.0
scipy==0.19.0
